post

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data= [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+                ];
+               
+        }
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        return view('tasks.index', $data);
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -49,12 +58,14 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
         
-        return redirect('/');
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+        
+        
+        return redirect('/tasks');
     }
 
     /**
@@ -65,11 +76,38 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
+        // そもそもログインしてなかったら、TOPにリダイレクト
+        // ログインしたユーザが投稿したタスクのページだったら、
+        // タスクを表示
+        // そうじゃなかったら、TOPにリダイレクト
+        if (\Auth::check()) {
+            // ログインしているユーザを取得
+            $user = \Auth::user();
+            // 表示したいタスクを取得
+            $task = Task::find($id);
+            
+            $userId = $user->id;
+            $taskUserId = $task->user_id;
+
+            // 表示したいタスクを作成したユーザのIDと、
+            // ログインしているユーザのIDが等しいか確認
+            if ($userId === $taskUserId) {
+                
+
+                $data = [
+                    'task' => $task,
+                ];
         
-        return view('tasks.show', [
-            'task' => $task,
-            ]);
+                return view('tasks.show', $data);
+
+            } else {
+                // TOPにリダイレクト
+                return redirect('/');
+            }
+        } else {
+            // ログインしてない場合TOPにリダイレクト
+            return redirect('/');
+        }
     }
 
     /**
@@ -80,11 +118,29 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $task = Task::find($id);
+            
+            $userId = $user->id;
+            $taskUserId = $task->user_id;
         
-        return view('tasks.edit', [
-            'task' => $task,
-            ]);
+            if ($userId === $taskUserId) {
+                
+
+                $data = [
+                    'task' => $task,
+                ];
+        
+                return view('tasks.edit', $data);
+
+                } else {
+                    return redirect('/');
+                }
+        } else {
+                return redirect('/');
+        }
+
     }
 
     /**
@@ -117,9 +173,31 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $task = \App\Task::find($id);
+            
+            $userId = $user->id;
+            $taskUserId = $task->user_id;
+        
+            if ($userId === $taskUserId) {
+                
+
+                $data = [
+                    'task' => $task,
+                ];
+                
+                 if (\Auth::id() === $task->user_id) {
+                    $task->delete();
+                   }
+                   return redirect('/');
+
+                } else {
+                    return redirect('/');
+                }
+        } else {
+                return redirect('/');
+        }
     }
 }
